Add Java 11 Support

+++ functions/java-jre.bash
@@ -3,7 +3,7 @@
 # shellcheck disable=SC2144
 
 ## Install appropriate Java version based on current choice.
-## Valid arguments: "AdoptOpenJDK", "Zulu32", or "Zulu64"
+## Valid arguments: "AdoptOpenJDK", "Zulu8-32", "Zulu8-64", "Zulu11-32", or "Zulu11-64"
 ##
 ##    java_install_and_update(String type)
 ##
@@ -18,17 +18,29 @@ java_install_or_update(){
         cond_redirect java_alternatives_reset "Zulu"
         cond_redirect systemctl start openhab2.service
       else
-        if [ "$1" == "Zulu64" ]; then
+        if [ "$1" == "Zulu8-64" ]; then
           if is_x86_64; then
-            java_zulu_enterprise_8_apt
+            java_zulu_enterprise_apt 8
           else
-            if java_zulu_tar_update_available 64-bit; then
-              java_zulu_8_tar 64-bit
+            if java_zulu_tar_update_available Zulu8-64; then
+              java_zulu_tar Zulu8-64
+            fi
+          fi
+        elif [ "$1" == "Zulu11-64" ]; then
+          if is_x86_64; then
+            java_zulu_enterprise_apt 11
+          else
+            if java_zulu_tar_update_available Zulu11-64; then
+              java_zulu_tar Zulu11-64
             fi
           fi
         else # Default to 32-bit installation
-          if java_zulu_tar_update_available 32-bit; then
-            java_zulu_8_tar 32-bit
+          if [ "$1" == "Zulu11-32" ]; then
+            if java_zulu_tar_update_available Zulu11-32; then
+              java_zulu_tar Zulu11-64
+            fi
+          elif java_zulu_tar_update_available Zulu8-32; then
+            java_zulu_tar Zulu8-32
           fi
         fi
       fi
@@ -37,12 +49,12 @@ java_install_or_update(){
   cond_redirect java -version
 }
 
-## Install Java Zulu 8 directly from fetched .tar.gz file
-## Valid arguments: "32-bit" or "64-bit"
+## Install Java Zulu directly from fetched .tar.gz file
+## Valid arguments: "Zulu8-32", "Zulu8-64", "Zulu11-32", or "Zulu11-64"
 ##
-##    java_zulu_8_tar(String arch)
+##    java_zulu_tar(String arch)
 ##
-java_zulu_8_tar(){
+java_zulu_tar(){
   local link
   local jdkTempLocation
   local jdkInstallLocation
@@ -51,9 +63,14 @@ java_zulu_8_tar(){
 
   cond_redirect systemctl stop openhab2.service
 
-  if [ "$1" == "32-bit" ]; then
-    echo -n "$(timestamp) [openHABian] Installing Java Zulu 32-Bit OpenJDK... "
-    link=$(fetch_zulu_tar_url 32-bit)
+  if [ "$1" == "Zulu8-32" ] || [ "$1" == "Zulu11-32" ]; then
+    if [ "$1" == "Zulu8-32" ]; then
+      echo -n "$(timestamp) [openHABian] Installing Java Zulu 8 32-Bit OpenJDK... "
+      link=$(fetch_zulu_tar_url Zulu8-32)
+    elif [ "$1" == "Zulu11-32" ]; then
+      echo -n "$(timestamp) [openHABian] Installing Java Zulu 11 32-Bit OpenJDK... "
+      link=$(fetch_zulu_tar_url Zulu11-32)
+    fi
     if is_arm; then
       jdkArch="aarch32"
     else
@@ -70,9 +87,14 @@ java_zulu_8_tar(){
       cond_redirect apt-get install --yes libc6:i386 libncurses5:i386 libstdc++6:i386
     fi
 
-  elif [ "$1" == "64-bit" ]; then
-    echo -n "$(timestamp) [openHABian] Installing Java Zulu 64-Bit OpenJDK... "
-    link=$(fetch_zulu_tar_url 64-bit)
+  elif [ "$1" == "Zulu8-64" ] || [ "$1" == "Zulu11-64" ]; then
+    if [ "$1" == "Zulu8-64" ]; then
+      echo -n "$(timestamp) [openHABian] Installing Java Zulu 8 64-Bit OpenJDK... "
+      link=$(fetch_zulu_tar_url Zulu8-64)
+    elif [ "$1" == "Zulu11-64" ]; then
+      echo -n "$(timestamp) [openHABian] Installing Java Zulu 11 64-Bit OpenJDK... "
+      link=$(fetch_zulu_tar_url Zulu11-64)
+    fi
     if is_aarch64 && [ "$(getconf LONG_BIT)" == "64" ]; then
       jdkArch="aarch64"
     elif is_x86_64 && [ "$(getconf LONG_BIT)" == "64" ]; then
@@ -81,9 +103,9 @@ java_zulu_8_tar(){
       if [ -n "$INTERACTIVE" ]; then
         whiptail --title "Incompatible hardware detected" --msgbox "Zulu OpenJDK 64-bit: this option does not currently work on your platform.\\n\\nDefaulting to 32-bit installation." 10 60
       else
-        echo "Zulu OpenJDK 64-bit: this option does not currently work on your platform. Defaulting to 32-bit installation."
+        echo "Zulu OpenJDK 64-bit: this option does not currently work on your platform. Defaulting to Java Zulu 8 32-bit installation."
       fi
-      link=$(fetch_zulu_tar_url 32-bit)
+      link=$(fetch_zulu_tar_url Zulu8-32)
       if is_arm; then
         jdkArch="aarch32"
       else
@@ -125,44 +147,77 @@ java_zulu_8_tar(){
 
 ## Install Azul's Java Zulu Enterprise using APT repository.
 ## (package manager distributions only available on x86-64bit platforms when checked in April 2020)
+## Valid arguments: "8" or "11"
 ##
-java_zulu_enterprise_8_apt(){
-  if ! dpkg -s 'zulu-8' > /dev/null 2>&1; then # Check if already is installed
-  echo -n "$(timestamp) [openHABian] Adding Zulu keys to apt... "
-  cond_redirect apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 219BD9C9
-  if [ $? -ne 0 ]; then echo "FAILED (keyserver)"; exit 1; fi
-  cond_redirect systemctl stop openhab2.service
-  echo -n "$(timestamp) [openHABian] Installing Zulu Enterprise 64-Bit OpenJDK... "
-  echo "deb http://repos.azulsystems.com/debian stable main" > /etc/apt/sources.list.d/zulu-enterprise.list
-  cond_redirect apt-get update
-  if cond_redirect apt-get --yes install zulu-8 && java_zulu_install_crypto_extension; then echo "OK"; else echo "FAILED"; exit 1; fi
-  cond_redirect systemctl start openhab2.service
-fi
+##    java_zulu_enterprise_apt(String ver)
+##
+java_zulu_enterprise_apt(){
+  if [ "$1" == "8" ]; then
+    if ! dpkg -s 'zulu-8' > /dev/null 2>&1; then # Check if already is installed
+      echo -n "$(timestamp) [openHABian] Adding Zulu keys to apt... "
+      cond_redirect apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 219BD9C9
+      if [ $? -ne 0 ]; then echo "FAILED (keyserver)"; exit 1; fi
+      cond_redirect systemctl stop openhab2.service
+      echo -n "$(timestamp) [openHABian] Installing Zulu 8 Enterprise 64-Bit OpenJDK... "
+      echo "deb http://repos.azulsystems.com/debian stable main" > /etc/apt/sources.list.d/zulu-enterprise.list
+      cond_redirect apt-get update
+      if cond_redirect apt-get --yes install zulu-8 && java_zulu_install_crypto_extension; then echo "OK"; else echo "FAILED"; exit 1; fi
+      cond_redirect systemctl start openhab2.service
+    fi
+  elif [ "$1" == "11" ]; then
+    if ! dpkg -s 'zulu-11' > /dev/null 2>&1; then # Check if already is installed
+      echo -n "$(timestamp) [openHABian] Adding Zulu keys to apt... "
+      cond_redirect apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys 219BD9C9
+      if [ $? -ne 0 ]; then echo "FAILED (keyserver)"; exit 1; fi
+      cond_redirect systemctl stop openhab2.service
+      echo -n "$(timestamp) [openHABian] Installing Zulu 11 Enterprise 64-Bit OpenJDK... "
+      echo "deb http://repos.azulsystems.com/debian stable main" > /etc/apt/sources.list.d/zulu-enterprise.list
+      cond_redirect apt-get update
+      if cond_redirect apt-get --yes install zulu-11 && java_zulu_install_crypto_extension; then echo "OK"; else echo "FAILED"; exit 1; fi
+      cond_redirect systemctl start openhab2.service
+    fi
+  fi
 }
 
 ## Fetch Java Zulu 8 directly from Azul API v1
-## Valid arguments: "32-bit" or "64-bit"
+## Valid arguments: "Zulu8-32", "Zulu8-64", "Zulu11-32", or "Zulu11-64"
 ##
 ##    fetch_zulu_tar_url(String arch)
 ##
 fetch_zulu_tar_url(){
-  local link
+  local link8
+  local link11
   local downloadlink
 
-  link="https://api.azul.com/zulu/download/community/v1.0/bundles/latest/binary/?jdk_version=8&ext=tar.gz&os=linux"
+  link8="https://api.azul.com/zulu/download/community/v1.0/bundles/latest/binary/?jdk_version=8&ext=tar.gz&os=linux"
+  link11="https://api.azul.com/zulu/download/community/v1.0/bundles/latest/binary/?jdk_version=11&ext=tar.gz&os=linux"
 
-  if [ "$1" == "32-bit" ]; then
+  if [ "$1" == "Zulu8-32" ]; then
     if is_arm; then
-      downloadlink=$(curl "${link}&arch=arm&hw_bitness=32" -s -L -I -o /dev/null -w '%{url_effective}' | sed -e 's#32sf#32hf#g')
+      downloadlink=$(curl "${link8}&arch=arm&hw_bitness=32" -s -L -I -o /dev/null -w '%{url_effective}' | sed -e 's#32sf#32hf#g')
     else
-      downloadlink=$(curl "${link}&arch=x86&hw_bitness=32" -s -L -I -o /dev/null -w '%{url_effective}' | sed -e 's#32sf#32hf#g')
+      downloadlink=$(curl "${link8}&arch=x86&hw_bitness=32" -s -L -I -o /dev/null -w '%{url_effective}' | sed -e 's#32sf#32hf#g')
     fi
 
-  elif [ "$1" == "64-bit" ]; then
+  elif [ "$1" == "Zulu11-32" ]; then
     if is_arm; then
-      downloadlink=$(curl "${link}&arch=arm&hw_bitness=64" -s -L -I -o /dev/null -w '%{url_effective}')
+      downloadlink=$(curl "${link11}&arch=arm&hw_bitness=32" -s -L -I -o /dev/null -w '%{url_effective}' | sed -e 's#32sf#32hf#g')
     else
-      downloadlink=$(curl "${link}&arch=x86&hw_bitness=64" -s -L -I -o /dev/null -w '%{url_effective}')
+      downloadlink=$(curl "${link11}&arch=x86&hw_bitness=32" -s -L -I -o /dev/null -w '%{url_effective}' | sed -e 's#32sf#32hf#g')
+    fi
+
+  elif [ "$1" == "Zulu8-64" ]; then
+    if is_arm; then
+      downloadlink=$(curl "${link8}&arch=arm&hw_bitness=64" -s -L -I -o /dev/null -w '%{url_effective}')
+    else
+      downloadlink=$(curl "${link8}&arch=x86&hw_bitness=64" -s -L -I -o /dev/null -w '%{url_effective}')
+    fi
+
+  elif [ "$1" == "Zulu11-64" ]; then
+    if is_arm; then
+      downloadlink=$(curl "${link11}&arch=arm&hw_bitness=64" -s -L -I -o /dev/null -w '%{url_effective}')
+    else
+      downloadlink=$(curl "${link11}&arch=x86&hw_bitness=64" -s -L -I -o /dev/null -w '%{url_effective}')
     fi
 
   else
@@ -177,7 +232,7 @@ fetch_zulu_tar_url(){
 
 ## Check if a newer version of Java Zulu 8 is available.
 ## Returns 0 / true if new version exists
-## Valid arguments: "32-bit" or "64-bit"
+## Valid arguments: "Zulu8-32", "Zulu8-64", "Zulu11-32", or "Zulu11-64"
 ##
 ##    java_zulu_tar_update_available(String arch)
 ##
@@ -186,27 +241,43 @@ java_zulu_tar_update_available(){
   local availableVersion
   local javaVersion
   local filter
-  local link
+  local link8
+  local link11
   if [ ! -x "$(command -v jq)" ]; then
     cond_redirect apt-get install --yes jq
   fi
 
   filter='[.zulu_version[] | tostring] | join(".")'
-  link="https://api.azul.com/zulu/download/community/v1.0/bundles/latest/?jdk_version=8&ext=tar.gz&os=linux"
+  link8="https://api.azul.com/zulu/download/community/v1.0/bundles/latest/?jdk_version=8&ext=tar.gz&os=linux"
+  link11="https://api.azul.com/zulu/download/community/v1.0/bundles/latest/?jdk_version=11&ext=tar.gz&os=linux"
   javaVersion=$(java -version 2>&1 | grep -m 1 -o "[0-9]\{0,3\}\.[0-9]\{0,3\}\.[0-9]\{0,3\}\.[0-9]\{0,3\}")
 
-  if [ "$1" == "32-bit" ]; then
+  if [ "$1" == "Zulu8-32" ]; then
     if is_arm; then
-      availableVersion=$(curl -s -H "Accept: application/json" "${link}&arch=arm&hw_bitness=32" | jq -r "$filter")
+      availableVersion=$(curl -s -H "Accept: application/json" "${link8}&arch=arm&hw_bitness=32" | jq -r "$filter")
     else
-      availableVersion=$(curl -s -H "Accept: application/json" "${link}&arch=x86&hw_bitness=32" | jq -r "$filter")
+      availableVersion=$(curl -s -H "Accept: application/json" "${link8}&arch=x86&hw_bitness=32" | jq -r "$filter")
     fi
 
-  elif [ "$1" == "64-bit" ]; then
+  elif [ "$1" == "Zulu11-32" ]; then
     if is_arm; then
-      availableVersion=$(curl -s -H "Accept: application/json" "${link}&arch=arm&hw_bitness=64" | jq -r "$filter")
+      availableVersion=$(curl -s -H "Accept: application/json" "${link11}&arch=arm&hw_bitness=32" | jq -r "$filter")
     else
-      availableVersion=$(curl -s -H "Accept: application/json" "${link}&arch=x86&hw_bitness=64" | jq -r "$filter")
+      availableVersion=$(curl -s -H "Accept: application/json" "${link11}&arch=x86&hw_bitness=32" | jq -r "$filter")
+    fi
+
+  elif [ "$1" == "Zulu8-64" ]; then
+    if is_arm; then
+      availableVersion=$(curl -s -H "Accept: application/json" "${link8}&arch=arm&hw_bitness=64" | jq -r "$filter")
+    else
+      availableVersion=$(curl -s -H "Accept: application/json" "${link8}&arch=x86&hw_bitness=64" | jq -r "$filter")
+    fi
+
+  elif [ "$1" == "Zulu11-64" ]; then
+    if is_arm; then
+      availableVersion=$(curl -s -H "Accept: application/json" "${link11}&arch=arm&hw_bitness=64" | jq -r "$filter")
+    else
+      availableVersion=$(curl -s -H "Accept: application/json" "${link11}&arch=x86&hw_bitness=64" | jq -r "$filter")
     fi
 
   else
@@ -255,23 +326,24 @@ adoptopenjdk_fetch_apt(){
   echo -n "$(timestamp) [openHABian] Fetching AdoptOpenJDK... "
   echo "deb http://adoptopenjdk.jfrog.io/adoptopenjdk/deb buster main" > /etc/apt/sources.list.d/adoptopenjdk.list
   cond_redirect apt-get update
-  if cond_redirect apt-get install --download-only adoptopenjdk-8-hotspot-jre; then echo "OK"; else echo "FAILED"; exit 1; fi
+  if cond_redirect apt-get install --download-only adoptopenjdk-11-hotspot-jre; then echo "OK"; else echo "FAILED"; exit 1; fi
 }
 
 ## Install AdoptOpenJDK using APT repository.
 ##
 adoptopenjdk_install_apt(){
-  if ! dpkg -s 'adoptopenjdk-8-hotspot-jre' > /dev/null 2>&1; then # Check if already is installed
-    echo -n "$(timestamp) [openHABian] Installing AdoptOpenJDK... "
+  if ! dpkg -s 'adoptopenjdk-11-hotspot-jre' > /dev/null 2>&1; then # Check if already is installed
+    echo -n "$(timestamp) [openHABian] Installing AdoptOpenJDK 11... "
     adoptopenjdk_fetch_apt
     cond_redirect systemctl stop openhab2.service
     cond_redirect java_alternatives_reset
-    if cond_redirect apt-get install --yes adoptopenjdk-8-hotspot-jre; then echo "OK"; else echo "FAILED"; exit 1; fi
+    if cond_redirect apt-get install --yes adoptopenjdk-11-hotspot-jre; then echo "OK"; else echo "FAILED"; exit 1; fi
     cond_redirect systemctl start openhab2.service
-  elif dpkg -s 'adoptopenjdk-8-hotspot-jre' >/dev/null 2>&1; then
+  elif dpkg -s 'adoptopenjdk-11-hotspot-jre' >/dev/null 2>&1; then
+    echo -n "$(timestamp) [openHABian] Reconfiguring AdoptOpenJDK 11... "
     cond_redirect systemctl stop openhab2.service
     cond_redirect java_alternatives_reset
-    if cond_redirect dpkg-reconfigure adoptopenjdk-8-hotspot-jre; then echo "OK"; else echo "FAILED"; exit 1; fi
+    if cond_redirect dpkg-reconfigure adoptopenjdk-11-hotspot-jre; then echo "OK"; else echo "FAILED"; exit 1; fi
     cond_redirect systemctl start openhab2.service
   fi
 }
